Electric Bike In PAkistan
An electric bike, often referred to as an e-bike, is a modern and eco-friendly mode of transportation that combines the benefits of a traditional bicycle with the convenience and assistance of an electric motor. These bicycles are equipped with a rechargeable battery-powered motor that assists the rider in pedaling, providing a smoother and less strenuous cycling experience.